Closes #106: move MotionEvent.isWithinBoundsOf and make it public
